refactor: Make endpoints and api module private
