+++ src/ikea_api/_endpoints/__init__.py
@@ -1,12 +1,12 @@
 from __future__ import annotations
 
-from ikea_api.endpoints.auth import get_authorized_token, get_guest_token
-from ikea_api.endpoints.cart import Cart
-from ikea_api.endpoints.item.item_ingka import fetch as _fetch_items_specs_ingka
-from ikea_api.endpoints.item.item_iows import fetch as _fetch_items_specs_iows
-from ikea_api.endpoints.item.item_pip import fetch as _fetch_items_specs_pip
-from ikea_api.endpoints.order_capture import OrderCapture
-from ikea_api.endpoints.purchases import Purchases
+from ikea_api._endpoints.auth import get_authorized_token, get_guest_token
+from ikea_api._endpoints.cart import Cart
+from ikea_api._endpoints.item.item_ingka import fetch as _fetch_items_specs_ingka
+from ikea_api._endpoints.item.item_iows import fetch as _fetch_items_specs_iows
+from ikea_api._endpoints.item.item_pip import fetch as _fetch_items_specs_pip
+from ikea_api._endpoints.order_capture import OrderCapture
+from ikea_api._endpoints.purchases import Purchases
 
 
 class Items:
