+++ src/ikea_api/_endpoints/cart.py
@@ -2,9 +2,9 @@ from __future__ import annotations
 
 from typing import Any
 
-from ikea_api.api import API, GraphQLResponse
+from ikea_api._api import API, GraphQLResponse
+from ikea_api._endpoints.item import parse_item_code
 from ikea_api.constants import Constants, Secrets
-from ikea_api.endpoints.item import parse_item_code
 
 
 class Cart(API):
