+++ src/ikea_api/_endpoints/item/item_pip.py
@@ -7,8 +7,8 @@ from typing import Any
 
 import aiohttp
 
+from ikea_api._endpoints.item import build_headers, parse_item_code
 from ikea_api.constants import Constants
-from ikea_api.endpoints.item import build_headers, parse_item_code
 from ikea_api.errors import ItemFetchError
